+++ src/swap/central/swapuz.ts
@@ -21,6 +21,7 @@ import {
 import { div18 } from '../../util/biggystringplus'
 import {
   checkInvalidCodes,
+  checkWhitelistedMainnetCodes,
   ensureInFuture,
   getCodesWithTranscription,
   getMaxSwappable,
@@ -65,8 +66,61 @@ const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
 // Network names that don't match parent network currency code
 // See https://swapuz.com/ for list of supported currencies
 const MAINNET_CODE_TRANSCRIPTION = {
+  arbitrum: 'ARBITRUM',
+  algorand: 'ALGO',
+  // avalanche:
+  // axelar:
+  // base:
+  // binance:
   binancesmartchain: 'BSC',
-  optimism: 'OPTIMISM'
+  bitcoin: 'BTC',
+  bitcoincash: 'BCH',
+  // bitcoingold:
+  // bitcoinsv:
+  celo: 'CELO',
+  // coreum:
+  // cosmoshub:
+  dash: 'DASH',
+  // digibyte:
+  dogecoin: 'DOGE',
+  // eboost:
+  eos: 'EOS',
+  ethereum: 'ETH',
+  ethereumclassic: 'ETC',
+  // ethereumpow:
+  fantom: 'FTM',
+  // feathercoin:
+  filecoin: 'FIL',
+  // filecoinfevm:
+  // fio:
+  // groestlcoin:
+  hedera: 'HBAR',
+  // liberland:
+  litecoin: 'LTC',
+  monero: 'XMR',
+  optimism: 'OPTIMISM',
+  // osmosis:
+  // piratechain:
+  polkadot: 'DOT',
+  polygon: 'MATIC',
+  // pulsechain:
+  qtum: 'QTUM',
+  ravencoin: 'RVN',
+  ripple: 'XRP',
+  // rsk:
+  // smartcash:
+  solana: 'SOL',
+  stellar: 'XLM',
+  // telos:
+  tezos: 'XTZ',
+  // thorchainrune:
+  tron: 'TRX',
+  // ufo:
+  // vertcoin:
+  // wax:
+  zcash: 'ZEC'
+  // zcoin:
+  // zksync:
 }
 
 export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
@@ -85,6 +139,7 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     const { fromWallet, toWallet } = request
 
     checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+    checkWhitelistedMainnetCodes(MAINNET_CODE_TRANSCRIPTION, request, swapInfo)
 
     // Grab addresses:
     const [fromAddress, toAddress] = await Promise.all([