+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: Mainnet codes for Arbitrum, Axelar, Base, and Cosmos Hub
+- changed: (CEX) Only allow quotes with known mainnet network codes
+
 ## 2.0.2 (2023-02-01)
 
 - fixed: Do not crash at load time if `BigInt` is not present.